backend/requirements.txt
# Flask Backend Requirements for CSS Picker SaaS

Flask==3.0.0
Flask-CORS==4.0.0
stripe==7.8.0
libsql==0.1.10
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0